+++ src/main/java/com/sms/springboot/service/StudentCourseService.java
@@ -10,8 +10,12 @@ import com.sms.springboot.model.StudentCourse;
 @Component
 public interface StudentCourseService {
 	List<StudentCourse> getAllStudentCourses();
+
 	StudentCourse saveStudentCourse(StudentCourse studentCourse);
-	StudentCourse getStudentCourseById(Long id);
+
+	StudentCourse getStudentCourseById(Integer id);
+
 	StudentCourse updateStudentCourse(StudentCourse studentCourse);
+
 	List<StudentCourse> getStudentCourseByStudent(Student student);
 }
